Allow for local file saves and task tests.
