+++ lib/worker.js
@@ -7,8 +7,7 @@ var assert = require("assert"),
 
 var async = require("async"),
     holdtime = require("holdtime"),
-    tmp = require("tmp"),
-    Uploader = require("s3-streaming-upload").Uploader;
+    tmp = require("tmp");
 
 // bump up the number of concurrent uploads for s3-streaming-upload
 // TODO it would be better to provide an agent, but that's not currently
@@ -27,7 +26,7 @@ https.globalAgent.maxSockets = Infinity;
  * args: []
  */
 
-module.exports = function(id, task, done) {
+module.exports = function(id, task, upload, done) {
   console.log("Processing task:", task.name);
 
   var outputURI;
@@ -37,10 +36,6 @@ module.exports = function(id, task, done) {
     assert.equal(1, task.operations.length, "Only single operation pipelines are supported at the moment");
     assert.ok(task.input, "An input is required");
     assert.ok(task.output, "An output is required (for now)");
-
-    outputURI = url.parse(task.output);
-
-    assert.equal("s3:", outputURI.protocol, "Only S3 outputs are supported");
   } catch (err) {
     return done(err);
   }
@@ -82,31 +77,16 @@ module.exports = function(id, task, done) {
       return done(err);
     }
 
-    // grab the last output
-    var output = outputs.pop();
-
-    // TODO tmp files aren't cleaned up when the process is terminated
-
     var cleanup = holdtime(function(err) {
       console.log("upload: %dms", arguments[arguments.length - 1]);
 
-      fs.unlink(output, function() {});
+      outputs.forEach(function(output) {
+        fs.unlink(output, function() {});
+      });
 
       return done(err);
     });
 
-    var uploader = new Uploader({
-      bucket: outputURI.hostname,
-      // strip the leading /
-      objectName: outputURI.pathname.slice(1),
-      objectParams: {
-        ACL: "public-read"
-      },
-      stream: fs.createReadStream(output)
-    });
-
-    uploader
-      .on("completed", cleanup)
-      .on("failed", cleanup);
+    upload(task.output, outputs, cleanup);
   });
 };
